Create check_coordinates.py
This code checks that coordinates from the telescope match the coordinates from the object table on ARI while taking proper motion into account. You just have to enter the path to the files to check this. This could still be implemented more broadly into one of the checks on ARI.